Codeforces DP page 3 done!

+++ Codeforces/Problemset/Combinatorics/1293b.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 using I=int;
-using Lf=long long int;
+using Lf=long double;
 
 I main(){
   cin.tie(0)->sync_with_stdio(0);
